python-server/requirements.txt
fastapi
opencv-python
numpy
requests
gradio
pillow
uvicorn
